--- a/src/bundle.c
+++ b/src/bundle.c
@@ -824,7 +824,7 @@ static int install_bundles(struct list *bundles, struct list **subs, int current
 	(void)rm_staging_dir_contents("download");
 
 download_subscribed_packs:
-	if (list_longer_than(to_install_files, 10) && download_subscribed_packs(*subs, mom, true, true)) {
+	if (list_longer_than(to_install_files, 10) && download_subscribed_packs(*subs, mom, true)) {
 		if (retries < MAX_TRIES && !content_url_is_local) {
 			increment_retries(&retries, &timeout);
 			printf("\nRetry #%d downloading subscribed packs\n", retries);