--- a/src/verify.c
+++ b/src/verify.c
@@ -323,7 +323,7 @@ static int get_all_files(struct manifest *official_manifest, struct list *subs)
 	struct list *iter;
 
 	/* for install we need everything so synchronously download zero packs */
-	ret = download_subscribed_packs(subs, official_manifest, true, false);
+	ret = download_subscribed_packs(subs, official_manifest, true);
 	if (ret < 0) { // require zero pack
 		/* If we hit this point, we know we have a network connection, therefore
 		 * 	the error is server-side. This is also a critical error, so detailed