--- a/test/functional/update/slow-server/lines-checked
+++ b/test/functional/update/slow-server/lines-checked
@@ -1,6 +1,10 @@
 Update started.
 Preparing to update from 10 to 100
+REGEXP:Error for .*/update/slow-server/state/pack-os-core-from-10-to-100.tar download: Response 206 - No error
+REGEXP:Starting download retry #1 for .*100/pack-os-core-from-10.tar
+REGEXP:Error for .*/slow-server/state/pack-os-core-from-10-to-100.tar download: Response 200 - Requested range was not delivered by the server
 Range command not supported by server, download resume disabled.
+REGEXP:Starting download retry #2 for .*100/pack-os-core-from-10.tar
 Extracting os-core pack for version 100
 Statistics for going from version 10 to version 100:
     changed bundles   : 1