Using parallel download API do download packs in parallel.

Closes #298