@matthewrsj 2018-07-20
For enhancements such as this please include some sort of analysis of the performance improvements in the PR itself.

I performed some very limited testing on my system updating from 23820 -> 23830 with the following bundles installed:
```
ansible            dev-utils-dev  kvm-host           os-core-update-dev     storage-utils-dev
bootloader         editors        libX11client       os-core-update-index*  sudo
c-basic            editors-dev    mail-utils         os-installer           sysadmin-basic
cockpit            emacs          mail-utils-dev     p11-kit                sysadmin-basic-dev
containers-basic   ethtool        mixer              package-utils          sysadmin-remote
desktop            git            network-basic      perl-basic             telemetrics
desktop-apps       go-basic       network-basic-dev  perl-basic-dev         user-basic
desktop-assets     go-basic-dev   openssh-server     python2-basic          vim
desktop-autostart  iproute2       openssl            python3-basic          web-server-basic
desktop-gnomelibs  joe            os-clr-on-clr      python-basic           which
desktop-i3         jq             os-core            redis-native           wpa_supplicant
desktop-locales    kernel-native  os-core-dev        shells                 xemacs
dev-utils          koji           os-core-update     storage-utils
```
I performed five updates each with the current client and the client from this PR

| current | new |
| --- | --- |
| 18.8s | 16.0s |
| 19.8s | 17.0s |
| 18.6s | 16.8s |
| 22.3s | 17.0s |
| 17.7s | 16.8s |
| **19.4s avg** | **16.7s avg** |

So ~14% speedup from this very limited test. Very nice. Clearly most of the time taken is in the pack extraction phase though.