@matthewrsj 2018-07-20
This appears to work with `mixin` integration as well. Thank you very much for ensuring that @otaviobp