@aaronovz1 2018-07-20
@matthewrsj if pack extraction is a bottleneck would there be benefit in doing them in parallel as well with some kind of async task processor?