@matthewrsj 2018-07-20
@aaronovz1 pack extraction is actually done via a callback on the download itself, so it should be happening asynchronously already. Correct @otaviobp?