@otaviobp 2018-07-20
@aaronovz1 @matthewrsj Downloads are performed in parallel by curl. Extraction are executed in the main thread and are executed one by one. The major reason we see gain in performance is because we keep downloading files while we are extracting a pack