@matthewrsj 2018-07-20
> The major reason we see gain in performance is because we keep downloading files while we are extracting a pack

Right, I meant asynchronously to pack downloads, not each other.