@aaronovz1 2018-07-20
The benefit might not outweigh the complexity but it does seem that while you would be downloading multiple files at once, the extraction is still done in series and would benefit from parallel processing. Otherwise you potentially have multiple files downloaded sitting there waiting for extraction. Obviously not something for this PR/ticket but figured it was worth mentioning as a future improvement.