@otaviobp 2018-07-25
V2:
 - not printing warning for pack downloads.
 - rebased on top of new master and fixing broken tests