@matthewrsj 2018-08-01
With this PR test/functional/update/slow-server took about 5 minutes to complete on my system. What's the reason for that? It wasn't super obvious to me looking at the changes you made to the slow-server code.