@otaviobp 2018-08-15
I couldn't reproduce this problem in update/slow-server test, but the add-rc test was really taking a really long time because of unnecessary retries.

 I improved the commit message explaining some changes I made in slow-server. There's a slow down because of 2 retries that are necessary to check if download resume is working. This slows down the test in 30s, so it shouldn't take 5 minutes to complete. It's taking less than 50s in my machine. All tests are running in around 6~7 minutes.

Can you retry using this new branch?