@matthewrsj 2018-08-15
[offline convo] @otaviobp to skip the long-running test if not in travis environment.