@otaviobp 2018-08-15
New version. Spiking slow-server test when not in travis