@otaviobp 2018-08-15
new version: Fixed proposed change and improved docs.