@otaviobp 2018-08-21
Fixed problems reported (other than the test using old lib) in new version